Add support for message filters
Unlike strongly-typed request filters that produce a response,
message filters have access to all JsonRpcMessage and do not
need to send a response but can via the McpServer on the MessageContext

Here's the prompt I gave to Claude Opus 4.5

Similar to how McpServerHandler includes NotificationHandlers that allow you to interecept an artbitrary notifications in addition to the request handlers. We should add a new property to McpServerFilters that allows developers to intercept arbitrary JsonRpcMessages.

To support this we should add a new McpMessageHandler and McpMessageFilter delegates which is a delegate Task McpMessageHandler(MessageContext message, CancellationToken cancellationToken) and delegate McpMessageHandler McpMessageFilter(McpMessageHandler next) It should be able to interecept all JsonRpcMessages, not just task. MessageContext should be the new base class for RequestContext<TParams>, that includes everything currently in RequestContext<TParams> asside from the Params, MatchPrimitives, and JsonRpcRequest properties. There should be a JsonRpcMessage property though which should be equivalent to RequestContext.JsonRpcRequest, so RequestContext will have both a JsonRpcMessage and JsonRpcRequest property.

Make sure this is well tested as the other filters are, and that you can either skip the default handlers or call them by calling next.